Fix logout page hydration by splitting client component
